I did some changes...
 - fixed Machine bugs where the output got duplicated
 - added Toaster and Toast
 - Hamburgers now need Toast instead of Bread
 - better navigation inside of the Cookie Cook Book and Vanilla Recipe
Book
 - added Hermit Plushies ( Members of HermitCraft in small :D )
 - added Material Extractor (to get resources that were creative-only
before
 - and more...

+++ src/me/gamma/cookies/util/MathHelper.java
@@ -4,11 +4,11 @@ package me.gamma.cookies.util;
 
 public class MathHelper {
 
-	public static int intpow(int b, int e) {
+	public static double intpow(double b, int e) {
 		if(e == 0) {
 			return 1;
 		} else {
-			int a = intpow(b, e / 2);
+			double a = intpow(b, e / 2);
 			a *= a;
 			if(e % 2 == 1)
 				a *= b;
